docker/flask-transcription/requirements.txt
flask
flask-cors
openai-whisper
torch
requests
ffmpeg-python
